+++ newsfragments/+ec3f5bf9.misc.rst
@@ -1,0 +1,1 @@
+Refactor code a bit to keep client and process fixtures separately, and extract config onto separate subpackage.
