+++ pyproject.toml
@@ -29,7 +29,7 @@ classifiers = [
 ]
 dependencies = [
     "pytest",
-    "port-for>=0.6.0",
+    "port-for>=0.7.1",
     "mirakuru",
     "boto3",
     "boto3-stubs[dynamodb]",
